Extract saveTinybirdToken helper and fix config path resolution
- Create shared src/cli/env.ts with saveTinybirdToken() helper
- Handle missing newline when appending to .env.local
- Use findConfigFile instead of getConfigPath in dev.ts
- Save .env.local in same directory as tinybird.json
